Removed startup class for .net6 app style + unused using removed
