+++ SwipeNFT.Infrastructure/CommandHandlers/Authentication/DisableUserCommandHandler.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Configuration;
 using SwipeNFT.Contracts.Request.Command.Authentication;
