+++ SwipeNFT.Infrastructure/CommandHandlers/Authentication/EnableUserCommandHandler.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Configuration;
 using SwipeNFT.Contracts.Request.Command.Authentication;
